pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "joblib>=1.5.1",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "psycopg2-binary>=2.9.10",
    "python-binance>=1.0.29",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.0",
    "sqlalchemy>=2.0.41",
    "ta>=0.11.0",
    "websockets>=15.0.1",
    "werkzeug>=3.1.3",
]
